Introduce separate package for execution plan
This is the code that does not depend on tasks.
