+++ subprojects/core/src/main/java/org/gradle/execution/SelectedTaskExecutionAction.java
@@ -22,6 +22,7 @@ import org.gradle.api.execution.TaskExecutionGraph;
 import org.gradle.api.execution.TaskExecutionGraphListener;
 import org.gradle.api.internal.GradleInternal;
 import org.gradle.api.internal.project.ProjectInternal;
+import org.gradle.execution.taskgraph.TaskExecutionGraphInternal;
 import org.gradle.internal.InternalListener;
 
 import java.util.Collection;
