+++ subprojects/core/src/main/java/org/gradle/execution/plan/DefaultExecutionPlan.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.execution.taskgraph;
+package org.gradle.execution.plan;
 
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Function;
