+++ subprojects/core/src/main/java/org/gradle/execution/plan/DefaultPlanExecutor.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.execution.taskgraph;
+package org.gradle.execution.plan;
 
 import org.gradle.api.Action;
 import org.gradle.api.NonNullApi;
