+++ subprojects/core/src/main/java/org/gradle/execution/plan/FailureCollector.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.execution.taskgraph;
+package org.gradle.execution.plan;
 
 import java.util.ArrayList;
 import java.util.List;
