+++ subprojects/core/src/main/java/org/gradle/execution/plan/LocalTaskNode.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.execution.taskgraph;
+package org.gradle.execution.plan;
 
 import com.google.common.collect.ImmutableCollection;
 import org.gradle.api.Action;
