+++ subprojects/core/src/main/java/org/gradle/internal/service/scopes/GradleScopeServices.java
@@ -50,19 +50,19 @@ import org.gradle.execution.DryRunBuildExecutionAction;
 import org.gradle.execution.ExcludedTaskFilteringBuildConfigurationAction;
 import org.gradle.execution.ProjectConfigurer;
 import org.gradle.execution.SelectedTaskExecutionAction;
-import org.gradle.execution.TaskExecutionGraphInternal;
 import org.gradle.execution.TaskNameResolvingBuildConfigurationAction;
 import org.gradle.execution.TaskSelector;
 import org.gradle.execution.commandline.CommandLineTaskConfigurer;
 import org.gradle.execution.commandline.CommandLineTaskParser;
+import org.gradle.execution.plan.DependencyResolver;
+import org.gradle.execution.plan.LocalTaskNodeExecutor;
+import org.gradle.execution.plan.NodeExecutor;
+import org.gradle.execution.plan.PlanExecutor;
+import org.gradle.execution.plan.TaskDependencyResolver;
+import org.gradle.execution.plan.TaskNodeDependencyResolver;
+import org.gradle.execution.plan.TaskNodeFactory;
 import org.gradle.execution.taskgraph.DefaultTaskExecutionGraph;
-import org.gradle.execution.taskgraph.DependencyResolver;
-import org.gradle.execution.taskgraph.LocalTaskNodeExecutor;
-import org.gradle.execution.taskgraph.NodeExecutor;
-import org.gradle.execution.taskgraph.PlanExecutor;
-import org.gradle.execution.taskgraph.TaskDependencyResolver;
-import org.gradle.execution.taskgraph.TaskNodeFactory;
-import org.gradle.execution.taskgraph.TaskNodeDependencyResolver;
+import org.gradle.execution.taskgraph.TaskExecutionGraphInternal;
 import org.gradle.internal.Factory;
 import org.gradle.internal.cleanup.BuildOutputCleanupRegistry;
 import org.gradle.internal.cleanup.DefaultBuildOutputCleanupRegistry;
