+++ subprojects/core/src/main/java/org/gradle/internal/service/scopes/TaskExecutionServices.java
@@ -63,9 +63,9 @@ import org.gradle.cache.PersistentCache;
 import org.gradle.caching.internal.controller.BuildCacheController;
 import org.gradle.caching.internal.tasks.TaskCacheKeyCalculator;
 import org.gradle.caching.internal.tasks.TaskOutputCacheCommandFactory;
-import org.gradle.execution.TaskExecutionGraphInternal;
-import org.gradle.execution.taskgraph.DefaultPlanExecutor;
-import org.gradle.execution.taskgraph.PlanExecutor;
+import org.gradle.execution.plan.DefaultPlanExecutor;
+import org.gradle.execution.plan.PlanExecutor;
+import org.gradle.execution.taskgraph.TaskExecutionGraphInternal;
 import org.gradle.initialization.BuildCancellationToken;
 import org.gradle.internal.classloader.ClassLoaderHierarchyHasher;
 import org.gradle.internal.cleanup.BuildOutputCleanupRegistry;
