+++ subprojects/core/src/test/groovy/org/gradle/execution/plan/DefaultPlanExecutorTest.groovy
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.execution.taskgraph
+package org.gradle.execution.plan
 
 import org.gradle.api.Action
 import org.gradle.api.Project
