+++ subprojects/core/src/test/groovy/org/gradle/internal/service/scopes/GradleScopeServicesTest.groovy
@@ -29,9 +29,9 @@ import org.gradle.execution.BuildConfigurationActionExecuter
 import org.gradle.execution.BuildExecuter
 import org.gradle.execution.DefaultBuildExecuter
 import org.gradle.execution.ProjectConfigurer
-import org.gradle.execution.TaskExecutionGraphInternal
 import org.gradle.execution.TaskSelector
 import org.gradle.execution.taskgraph.DefaultTaskExecutionGraph
+import org.gradle.execution.taskgraph.TaskExecutionGraphInternal
 import org.gradle.initialization.BuildCancellationToken
 import org.gradle.internal.build.BuildState
 import org.gradle.internal.concurrent.DefaultParallelismConfiguration
