+++ subprojects/dependency-management/src/main/java/org/gradle/api/internal/artifacts/transform/TransformNode.java
@@ -29,8 +29,8 @@ import org.gradle.api.internal.artifacts.ivyservice.ResolvedArtifactCollectingVi
 import org.gradle.api.internal.artifacts.ivyservice.resolveengine.artifact.BuildableSingleResolvedArtifactSet;
 import org.gradle.api.internal.artifacts.ivyservice.resolveengine.artifact.ResolvableArtifact;
 import org.gradle.api.internal.artifacts.ivyservice.resolveengine.artifact.ResolvedArtifactSet;
-import org.gradle.execution.taskgraph.Node;
-import org.gradle.execution.taskgraph.TaskDependencyResolver;
+import org.gradle.execution.plan.Node;
+import org.gradle.execution.plan.TaskDependencyResolver;
 import org.gradle.internal.operations.BuildOperationCategory;
 import org.gradle.internal.operations.BuildOperationContext;
 import org.gradle.internal.operations.BuildOperationDescriptor;
